Fix types of new `MAX_*` contant values

+++ pallets/gear/src/benchmarking/mod.rs
@@ -108,9 +108,9 @@ use sp_runtime::{
 };
 use sp_std::{num::NonZero, prelude::*};
 
-const MAX_PAYLOAD_LEN: u32 = Payload::MAX_LEN;
+const MAX_PAYLOAD_LEN: u32 = Payload::MAX_LEN as u32;
 const MAX_PAYLOAD_LEN_KB: u32 = MAX_PAYLOAD_LEN / 1024;
-const MAX_SALT_SIZE_BYTES: u32 = Salt::MAX_LEN;
+const MAX_SALT_SIZE_BYTES: u32 = Salt::MAX_LEN as u32;
 const MAX_NUMBER_OF_DATA_SEGMENTS: u32 = 1024;
 const MAX_TABLE_ENTRIES: u32 = 10_000_000;
